Image enhancing, background removal, match_results saving to file
Image enhancing still needs some work, plan is to implement morphological opening to get rid of very tiny (1-3 pixels) fake peaks.

Background removal currently works that the user has to prepare a background array from a vacuum region averaged. To do: make the user choose a region of the sample within the function itself, average the patterns and then subtract automatically.

Save match_results to file - the format is so that it can be imported into MTEX matlab package for further orientation mapping.