Add ellptical disk and ring to MakeTestData class
They are added as two new classes, in addition to the original
Ring and Disk classes. Ideally, EllipseRing and Ring,
EllipseDisk and Disk, would be merged into a general class.
However, due to time constraints, this is not done this time.

Also add support in generate_4d_data for the new EllipseRing
functionality.