Major refactoring of ransac_ellipse_tools: use focus instead of centre
Essentially this commit changes how the is_ellipse_good checks the
ellipse model. It is changed from using the ellipse centre position,
to rather use the focus closest to the centre of the diffraction
pattern.

This change further necessitated _many_ changes elsewhere in the
ellipse model functions, including the markers, dummy data and unit
tests.