Created DiffractionPeaks class
1) Extension in the DiffractionProfile class:
Created the find_peaks_1D() funciton:
It takes a DiffractionProfile and finds peaks using the o'Haver function from Hyperspy. It returns the DiffractionPeaks class.
2) Added import class in pyxem __init__
3) Created the DiffractionPeaks class, the 1D equivalent of a DiffracionVectors class.

TO DO:
a) To expand DiffractionProfile class, to calibrate the axes to reciprocal space.
b) To store the intensity of each peak in the "intensity" attribute, with navigation dimensions identical to the original object.