Extension of DiffractionVectors class adding geometrical calibration function
1) Added the function calculate_detector_px_to_cartesian_diffraction_coordinates, which is similar to calculate_cartesian_coordinates, but it strarts from a signal in pixel units of the detector. Also, it is a purely geometrically derived function, which can be used for XRD (sharper Ewald spheres).
It calls the new sub-function detector_px_to_3D_kspace, which takes detector pixel units instead of k_x, k_y values.

TO DO:
a) Move the detector_px_to_1D_kspace function to vector_utils.py, as it is the geometrical equivalent of detector_to_fourier, but starting with pixel units in the detector.