Skip some ransac ellipse model unit tests
These were broken due to changes in skimage 0.19, thus this might also
break some of the functionality. The plan is to reimplement this
functionality in a future version of pyxem, utilizing the new
map functionality in HyperSpy. Thus, since these will be reimplemented
in the future, we're skipping them for the time being.