Update test_electron_diffraction2d.py
Changed set_diffraction_calibration docstring to reflect actual functionality. The previous code did not work as intended for already calibrated signals. Another keyword argument is added `center_pixels` that give the possibility to give the center of the diffraction pattern in pixel coordinates as well.

Made corresponding changes to relevant tests as well.