Update test_electron_diffraction2d.py
Changed set_diffraction_calibration docstring to reflect actual functionality. The previous code did not work as intended for already calibrated signals. Another keyword argument is added `pixels` that give the possibility to specify whether the center of the diffraction pattern is given in pixel coordinates or not.

Updated test_electron_diffraction2d to reflect changes in set_electron_diffraction().

Co-Authored-By: Carter Francis <csfrancis@wisc.edu>